Convert ConvDepthwiseFwd2D to dlopen and split CK utility header
Move the last remaining direct-CK solver (ConvDepthwiseFwd2D) behind
the dlopen boundary and split implicitgemm_ck_util.hpp so that no CK
headers leak into libMIOpen.so from converted solver files.

Header split:
- Extract CK-free utilities into implicitgemm_ck_util_common.hpp
  (workspace helpers, transpose classes, perf-config iteration)
- Move CK-dependent remainder to src/ck_impl/implicitgemm_ck_util.hpp
  (invoker factories, DeviceOp aliases, CK arg support checks)
- Replace ck::half_t/bhalf_t/tf32_t type tags in 3D solver AI
  heuristics with local CK-free tag structs

Depthwise conversion:
- New ck_depthwise_fwd_impl.cpp with 12 CK kernel instantiations
  and 5 extern "C" functions behind CK_IMPL_API_VERSION 5
- Solver stripped to use CkImplLibLoader for all CK operations
- Test updated with UsesCKDynamicLib() and corrected device list

Verified: libMIOpen.so contains zero ck:: symbols and zero device
code sections. All 244 tests pass across 9 test suites.
