+++ projects/miopen/src/ck_impl/CMakeLists.txt
@@ -23,6 +23,7 @@ set(CK_IMPL_SOURCES
     ck_fused_bias_res_add_activ_impl.cpp
     ck_fused_grp_activ_impl.cpp
     ck_fused_grp_bias_activ_impl.cpp
+    ck_depthwise_fwd_impl.cpp
 )
 
 # Strip target features (e.g. ":sramecc+:xnack-") to get the base arch name.
