+++ projects/miopen/src/ck_impl/ck_fused_grp_activ_impl.cpp
@@ -16,7 +16,7 @@
 #include <miopen/conv/data_invoke_params.hpp>
 #include <miopen/fusion/fusion_invoke_params.hpp>
 #include <miopen/solver/ck_utility_common.hpp>
-#include <miopen/solver/implicitgemm_ck_util.hpp>
+#include "implicitgemm_ck_util.hpp"
 #include <ck/library/tensor_operation_instance/gpu/grouped_convolution_forward_clamp.hpp>
 
 // ---------------------------------------------------------------------------
