+++ projects/miopen/src/ck_impl/ck_grouped_conv_3d_wrw_impl.cpp
@@ -15,7 +15,7 @@
 #include <miopen/execution_context.hpp>
 #include <miopen/conv/wrw_invoke_params.hpp>
 #include <miopen/solver/problem_description_interpreter.hpp>
-#include <miopen/solver/implicitgemm_ck_util.hpp>
+#include "implicitgemm_ck_util.hpp"
 
 namespace {
