+++ projects/miopen/src/include/miopen/conv/heuristics/ai_conv_3d_kernel_tuning_utils.hpp
@@ -34,7 +34,6 @@
 #include <miopen/conv/heuristics/ai_heuristics.hpp>
 #include <miopen/conv/heuristics/ai_candidate_selection.hpp>
 #include <miopen/execution_context.hpp>
-#include <miopen/solver/implicitgemm_ck_util.hpp>
 #include <miopen/solver/implicitgemm_util.hpp>
 
 #if MIOPEN_ENABLE_AI_KERNEL_TUNING
