+++ projects/miopen/src/include/miopen/solver/ck_impl_interface.hpp
@@ -28,7 +28,7 @@ struct ConvSolution;
 
 /// API version constant. Bump when the set of extern "C" symbols or their
 /// semantics change.  The loader checks this at dlopen time.
-#define CK_IMPL_API_VERSION 4
+#define CK_IMPL_API_VERSION 5
 
 /// Opaque handle wrapping a list of valid kernel ID strings.
 /// Allocated by the impl library, freed by the caller via
@@ -420,6 +420,40 @@ ck_impl_fused_grp_bias_activ_get_solution(const miopen::ExecutionContext* ctx,
                                             const char* kernel_id,
                                             bool use_tf32,
                                             miopen::solver::ConvSolution** out_solution);
+
+// -- Depthwise Conv FWD (custom CK kernel, FP16) ---------------------------------
+
+CK_IMPL_NODISCARD CK_IMPL_API ck_impl_status_t
+ck_impl_depthwise_fwd_fill_valid_kernels(const miopen::conv::ProblemDescription* problem,
+                                           miopenDataType_t data_type,
+                                           bool use_tf32,
+                                           CKKernelListHandle** out_handle);
+
+CK_IMPL_NODISCARD CK_IMPL_API ck_impl_status_t
+ck_impl_depthwise_fwd_is_applicable(const miopen::conv::ProblemDescription* problem,
+                                      miopenDataType_t data_type,
+                                      bool use_tf32,
+                                      bool* out_result);
+
+CK_IMPL_NODISCARD CK_IMPL_API ck_impl_status_t
+ck_impl_depthwise_fwd_is_args_supported(const miopen::conv::ProblemDescription* problem,
+                                          const char* kernel_id,
+                                          miopenDataType_t data_type,
+                                          bool use_tf32,
+                                          bool* out_result);
+
+CK_IMPL_NODISCARD CK_IMPL_API ck_impl_status_t
+ck_impl_depthwise_fwd_get_workspace_size(const miopen::conv::ProblemDescription* problem,
+                                           miopenDataType_t data_type,
+                                           bool use_tf32,
+                                           size_t* out_size);
+
+CK_IMPL_NODISCARD CK_IMPL_API ck_impl_status_t
+ck_impl_depthwise_fwd_get_solution(const miopen::ExecutionContext* ctx,
+                                     const miopen::conv::ProblemDescription* problem,
+                                     const char* kernel_id,
+                                     bool use_tf32,
+                                     miopen::solver::ConvSolution** out_solution);
 
 // -- Get all kernel type strings (for test/metadata validation) -----------------
