+++ projects/miopen/src/solver/conv/conv_hip_implicit_gemm_3d_grouped_fwd_xdlops.cpp
@@ -39,7 +39,7 @@
 #include <miopen/conv/heuristics/ai_candidate_selection.hpp>
 #include <miopen/conv/heuristics/ai_conv_3d_kernel_tuning_utils.hpp>
 #endif
-#include <miopen/solver/implicitgemm_ck_util.hpp>
+#include <miopen/solver/implicitgemm_ck_util_common.hpp>
 #include <miopen/solver/ck_impl_lib_loader.hpp>
 
 MIOPEN_DECLARE_ENV_VAR_BOOL(MIOPEN_DEBUG_3D_CONV_IMPLICIT_GEMM_HIP_FWD_XDLOPS)
@@ -50,6 +50,16 @@ MIOPEN_DECLARE_ENV_VAR_BOOL(MIOPEN_DEBUG_CK_DEFAULT_KERNELS)
 namespace miopen {
 namespace solver {
 namespace conv {
+
+namespace {
+// Type tags for AI heuristics. These replace CK types (ck::half_t, ck::bhalf_t,
+// ck::tf32_t) that were previously used as template arguments to
+// RunParameterPredictionModel. The DataType template parameter is unused in the
+// function body — these tags only drive the mode_use_tf32 constexpr check.
+struct HalfTag {};
+struct BFloat16Tag {};
+struct TF32Tag {};
+} // namespace
 
 using ProblemDescription = miopen::conv::ProblemDescription;
 
@@ -355,7 +365,7 @@ void PerformanceConfigHipImplicitGemm3DGroupFwdXdlops::HeuristicInit(
             using T        = decltype(CKDataType);
             using TCompute = decltype(CKComputeType);
             constexpr bool mode_use_tf32 =
-                std::is_same_v<T, float> && std::is_same_v<TCompute, ck::tf32_t>;
+                std::is_same_v<T, float> && std::is_same_v<TCompute, TF32Tag>;
 
             auto fill_valid_kernels =
                 [&loader](const ProblemDescription& p) -> std::vector<std::string> {
@@ -391,12 +401,12 @@ void PerformanceConfigHipImplicitGemm3DGroupFwdXdlops::HeuristicInit(
         switch(problem.GetInDataType())
         {
         case miopenHalf:
-            std::tie(ai_success, result) = run_ai_heuristics(ck::half_t{}, ck::half_t{});
+            std::tie(ai_success, result) = run_ai_heuristics(HalfTag{}, HalfTag{});
             break;
         case miopenFloat:
             if(problem.UseTF32())
             {
-                std::tie(ai_success, result) = run_ai_heuristics(float{}, ck::tf32_t{});
+                std::tie(ai_success, result) = run_ai_heuristics(float{}, TF32Tag{});
                 if(!ai_success || result.IsEmpty())
                 {
                     MIOPEN_LOG_I2("Step 3: AI heuristics with TF32 failed, retrying with FP32");
@@ -409,7 +419,7 @@ void PerformanceConfigHipImplicitGemm3DGroupFwdXdlops::HeuristicInit(
             }
             break;
         case miopenBFloat16:
-            std::tie(ai_success, result) = run_ai_heuristics(ck::bhalf_t{}, ck::bhalf_t{});
+            std::tie(ai_success, result) = run_ai_heuristics(BFloat16Tag{}, BFloat16Tag{});
             break;
         default: break;
         }
