+++ projects/miopen/src/solver/conv/conv_hip_implicit_gemm_3d_grouped_wrw_xdlops.cpp
@@ -46,12 +46,22 @@ MIOPEN_DECLARE_ENV_VAR_BOOL(MIOPEN_DEBUG_CK_DEFAULT_KERNELS)
 #include <miopen/conv/heuristics/ai_conv_3d_kernel_tuning_utils.hpp>
 #endif
 #endif
-#include <miopen/solver/implicitgemm_ck_util.hpp>
+#include <miopen/solver/implicitgemm_ck_util_common.hpp>
 #include <miopen/solver/implicitgemm_util.hpp>
 
 namespace miopen {
 namespace solver {
 namespace conv {
+
+namespace {
+// Type tags for AI heuristics. These replace CK types (ck::half_t, ck::bhalf_t,
+// ck::tf32_t) that were previously used as template arguments to
+// RunParameterPredictionModel. The DataType template parameter is unused in the
+// function body — these tags only drive the mode_use_tf32 constexpr check.
+struct HalfTag {};
+struct BFloat16Tag {};
+struct TF32Tag {};
+} // namespace
 
 using ProblemDescription = miopen::conv::ProblemDescription;
 
@@ -202,7 +212,7 @@ void PerformanceConfigHipImplicitGemm3DGroupWrwXdlops::HeuristicInit(
             using T        = decltype(CKDataType);
             using TCompute = decltype(CKComputeType);
             constexpr bool mode_use_tf32 =
-                std::is_same_v<T, float> && std::is_same_v<TCompute, ck::tf32_t>;
+                std::is_same_v<T, float> && std::is_same_v<TCompute, TF32Tag>;
             auto fill_valid_kernels =
                 [&loader](const ::miopen::conv::ProblemDescription& p) -> std::vector<std::string> {
                 return loader.FillValidKernels(
@@ -236,12 +246,12 @@ void PerformanceConfigHipImplicitGemm3DGroupWrwXdlops::HeuristicInit(
         switch(problem.GetInDataType())
         {
         case miopenHalf:
-            std::tie(ai_success, result) = run_ai_heuristics(ck::half_t{}, ck::half_t{});
+            std::tie(ai_success, result) = run_ai_heuristics(HalfTag{}, HalfTag{});
             break;
         case miopenFloat:
             if(problem.UseTF32())
             {
-                std::tie(ai_success, result) = run_ai_heuristics(float{}, ck::tf32_t{});
+                std::tie(ai_success, result) = run_ai_heuristics(float{}, TF32Tag{});
                 if(!ai_success || result.IsEmpty())
                 {
                     MIOPEN_LOG_I2("Step 3: AI heuristics with TF32 failed, retrying with FP32");
@@ -254,7 +264,7 @@ void PerformanceConfigHipImplicitGemm3DGroupWrwXdlops::HeuristicInit(
             }
             break;
         case miopenBFloat16:
-            std::tie(ai_success, result) = run_ai_heuristics(ck::bhalf_t{}, ck::bhalf_t{});
+            std::tie(ai_success, result) = run_ai_heuristics(BFloat16Tag{}, BFloat16Tag{});
             break;
         default: break;
         }
