+++ projects/miopen/test/gtest/unit_conv_solver_ConvCkGroupedConvFwd.cpp
@@ -120,12 +120,13 @@ auto GetTestParams(miopenDataType_t datatype)
 // Solution is only applicable to devices that default to 64-lane wavefronts
 #if MIOPEN_BACKEND_HIP && MIOPEN_USE_COMPOSABLEKERNEL
     Gpu supportedDevices =
-        Gpu::gfx900 | Gpu::gfx906 | Gpu::gfx908 | Gpu::gfx90A | Gpu::gfx94X | Gpu::gfx950;
+        Gpu::gfx908 | Gpu::gfx90A | Gpu::gfx94X | Gpu::gfx950;
 #else
     Gpu supportedDevices = Gpu::None;
 #endif
     auto params = miopen::unit_tests::UnitTestConvSolverParams(supportedDevices);
     params.Tunable(5);
+    params.UsesCKDynamicLib();
 
     return params;
 }
